# docs/Swagger_OpenAPI_Specs.json
{"openapi":"3.1.0","info":{"title":"Food Ordering Backend","version":"1.0.0"},"paths":{"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["Auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/restaurants/":{"get":{"tags":["Restaurants"],"summary":"List Restaurants","operationId":"list_restaurants_restaurants__get","parameters":[{"name":"last_id","in":"query","required":false,"schema":{"type":"integer","description":"Last ID of the previous page","default":0,"title":"Last Id"},"description":"Last ID of the previous page"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Size"}},{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search query for restaurant name","title":"Name"},"description":"Search query for restaurant name"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestaurantListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/restaurants/{restaurant_id}":{"get":{"tags":["Restaurants"],"summary":"Get Restaurant","operationId":"get_restaurant_restaurants__restaurant_id__get","parameters":[{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"integer","title":"Restaurant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestaurantResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/restaurants/registerRestaurant":{"post":{"tags":["Restaurants"],"summary":"Register Restaurant","operationId":"register_restaurant_restaurants_registerRestaurant_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestaurantCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestaurantResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/menu/restaurants/{restaurant_id}/menu":{"get":{"tags":["Menu"],"summary":"Get Menu","operationId":"get_menu_menu_restaurants__restaurant_id__menu_get","parameters":[{"name":"restaurant_id","in":"path","required":true,"schema":{"type":"integer","title":"Restaurant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MenuItemResponse"},"title":"Response Get Menu Menu Restaurants  Restaurant Id  Menu Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/menu/menu_items":{"post":{"tags":["Menu"],"summary":"Create Menu Item","operationId":"create_menu_item_menu_menu_items_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MenuItemCreate"},"type":"array","title":"Menu Item"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MenuItemResponse"},"type":"array","title":"Response Create Menu Item Menu Menu Items Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/orders/":{"post":{"tags":["Orders"],"summary":"Create Order","operationId":"create_order_orders__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Orders"],"summary":"List Orders","operationId":"list_orders_orders__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"last_id","in":"query","required":false,"schema":{"type":"integer","description":"Last ID of the previous page","default":0,"title":"Last Id"},"description":"Last ID of the previous page"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/{order_id}":{"get":{"tags":["Orders"],"summary":"Get Order","operationId":"get_order_orders__order_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MenuItemCreate":{"properties":{"restaurant_id":{"type":"integer","title":"Restaurant Id","description":"The ID of the restaurant to which the menu item belongs.","examples":[10,20,30]},"name":{"type":"string","title":"Name","description":"The name of the menu item.","examples":["Margherita Pizza","California Roll"]},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"A brief description of the menu item.","examples":["Classic pizza with tomatoes and mozzarella","Roll with crab and avocado"]},"price":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"}],"title":"Price","description":"The price of the menu item.","examples":["9.99","12.50"]}},"type":"object","required":["restaurant_id","name","description","price"],"title":"MenuItemCreate"},"MenuItemResponse":{"properties":{"id":{"type":"integer","title":"Id"},"restaurant_id":{"type":"integer","title":"Restaurant Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)","title":"Price"}},"type":"object","required":["id","restaurant_id","name","price"],"title":"MenuItemResponse"},"OrderCreate":{"properties":{"restaurant_id":{"type":"integer","title":"Restaurant Id","description":"The ID of the restaurant from which the order is placed.","examples":[10,20,30]},"order_items":{"items":{"$ref":"#/components/schemas/OrderItemCreate"},"type":"array","title":"Order Items","description":"A list of items included in the order."}},"type":"object","required":["restaurant_id","order_items"],"title":"OrderCreate"},"OrderItemCreate":{"properties":{"menu_item_id":{"type":"integer","title":"Menu Item Id","description":"The ID of the menu item being ordered.","examples":[100,200,300]},"quantity":{"type":"integer","maximum":1000.0,"exclusiveMinimum":0.0,"title":"Quantity","description":"The quantity of the menu item being ordered.","examples":[1,2,3]}},"type":"object","required":["menu_item_id","quantity"],"title":"OrderItemCreate"},"OrderItemResponse":{"properties":{"id":{"type":"integer","title":"Id"},"order_id":{"type":"integer","title":"Order Id"},"menu_item":{"$ref":"#/components/schemas/MenuItemResponse"},"quantity":{"type":"integer","maximum":1000.0,"exclusiveMinimum":0.0,"title":"Quantity"},"price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)","title":"Price"}},"type":"object","required":["id","order_id","menu_item","quantity","price"],"title":"OrderItemResponse"},"OrderListResponse":{"properties":{"total_counts":{"type":"integer","title":"Total Counts"},"last_id":{"type":"integer","title":"Last Id"},"size":{"type":"integer","title":"Size"},"orders":{"items":{"$ref":"#/components/schemas/OrderResponse"},"type":"array","title":"Orders"}},"type":"object","required":["total_counts","last_id","size","orders"],"title":"OrderListResponse"},"OrderResponse":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"restaurant_id":{"type":"integer","title":"Restaurant Id"},"total_amount":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)","title":"Total Amount"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"order_items":{"items":{"$ref":"#/components/schemas/OrderItemResponse"},"type":"array","title":"Order Items","default":[]}},"type":"object","required":["id","user_id","restaurant_id","total_amount","status","created_at"],"title":"OrderResponse"},"RestaurantCreate":{"properties":{"name":{"type":"string","title":"Name","description":"The name of the restaurant.","examples":["Pizza Palace","Sushi World"]},"address":{"type":"string","title":"Address","description":"The address of the restaurant.","examples":["123 Main Street","456 Oak Avenue"]},"phone_number":{"anyOf":[{"type":"string","maxLength":15,"minLength":10,"pattern":"^\\+?[1-9]\\d{1,14}$"},{"type":"null"}],"title":"Phone Number","description":"The contact phone number of the restaurant.","examples":["+1234567890","+1987654321"]}},"type":"object","required":["name","address","phone_number"],"title":"RestaurantCreate"},"RestaurantListResponse":{"properties":{"total_counts":{"type":"integer","title":"Total Counts"},"last_id":{"type":"integer","title":"Last Id"},"size":{"type":"integer","title":"Size"},"restaurants":{"items":{"$ref":"#/components/schemas/RestaurantResponse"},"type":"array","title":"Restaurants"}},"type":"object","required":["total_counts","last_id","size","restaurants"],"title":"RestaurantListResponse"},"RestaurantResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"address":{"type":"string","title":"Address"},"phone_number":{"anyOf":[{"type":"string","maxLength":15,"minLength":10,"pattern":"^\\+?[1-9]\\d{1,14}$"},{"type":"null"}],"title":"Phone Number"}},"type":"object","required":["id","name","address"],"title":"RestaurantResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserCreate":{"properties":{"email":{"type":"string","maxLength":254,"pattern":"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$","title":"Email","description":"The email address of the user.","examples":["user@example.com"]},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password","description":"The password for the user account.","examples":["StrongP@ssw0rd"]},"role":{"anyOf":[{"$ref":"#/components/schemas/UserRole","maxLength":50},{"type":"null"}],"default":"customer"}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserLogin":{"properties":{"email":{"type":"string","maxLength":254,"pattern":"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$","title":"Email","description":"The email address of the user.","examples":["The email address of the user."]},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password","description":"The password for the user account.","examples":["The password for the user account."]}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","maxLength":254,"pattern":"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$","title":"Email"},"role":{"$ref":"#/components/schemas/UserRole"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","email","role","created_at"],"title":"UserResponse"},"UserRole":{"type":"string","enum":["admin","customer","restaurant_owner"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}